latent_space_zoom/requirements.txt
torch
torchvision
diffusers
transformers
accelerate
Pillow
